requirements.txt
asgiref==3.7.2
Django==5.0.1
mysqlclient==2.2.1
sqlparse==0.4.4
tzdata==2023.4
fpdf2== 2.7.8
numlet==2.0.0